Move ConfirmationModalProps to types
